# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Comptable",
        "description": "API pour la gestion comptable avec journal des écritures et génération de balance comptable",
        "contact": {
            "name": "Support API",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentifier un utilisateur et générer un token",
                "description": "Authentifier un utilisateur et générer un token.",
                "operationId": "d618a255475240ba3d40cc39692cd924",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|laravel_sanctum_token"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/Utilisateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentification échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Identifiants invalides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnecter l'utilisateur",
                "description": "Déconnecter l'utilisateur (révoquer le token).",
                "operationId": "1e357eb2c93a1bdabebcebfa088332bf",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/export-balance": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "Exporter la balance comptable au format Excel",
                "description": "Générer un fichier Excel avec la balance comptable.",
                "operationId": "927ff6672295e78c3c4c4559bfad826e",
                "responses": {
                    "200": {
                        "description": "Fichier HTML permettant de télécharger la balance en Excel",
                        "content": {
                            "text/html": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/export-balance/csv": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "Exporter la balance comptable au format CSV",
                "description": "Exporter la balance comptable au format CSV.",
                "operationId": "a529b81b6e8171c8c461b316a165057a",
                "responses": {
                    "200": {
                        "description": "Fichier CSV de la balance comptable",
                        "content": {
                            "text/csv": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Utilisateur": {
                "title": "Utilisateur",
                "description": "Modèle d'utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+33612345678"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Rue de Paris"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "comptable",
                            "utilisateur"
                        ],
                        "example": "comptable"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}